+++ packages/sdks/e2e/vue/src/env.d.ts
@@ -6,5 +6,3 @@ declare module '*.vue' {
   const component: DefineComponent<{}, {}, any>;
   export default component;
 }
-
-declare module '@builder.io/sdk-vue';