+++ packages/sdks/src/blocks/columns/columns.lite.tsx
@@ -72,7 +72,9 @@ export default function Columns(props: PropsWithBuilderData<ColumnProps>) {
 
     get columnsCssVars(): Dictionary<string> {
       return useTarget({
-        reactNative: { flexDirection: state.flexDir },
+        reactNative: {
+          flexDirection: state.flexDir as 'row' | 'column' | 'column-reverse',
+        },
         default: {
           '--flex-dir': state.flexDir,
           '--flex-dir-tablet': state.getTabletStyle({
@@ -222,7 +224,12 @@ export default function Columns(props: PropsWithBuilderData<ColumnProps>) {
               })}
               path={`component.options.columns.${index}.blocks`}
               parent={props.builderBlock.id}
-              styleProp={{ flexGrow: '1' }}
+              styleProp={{
+                flexGrow: useTarget<string | number>({
+                  reactNative: 1,
+                  default: '1',
+                }),
+              }}
               context={props.builderContext}
               registeredComponents={props.builderComponents}
               linkComponent={props.builderLinkComponent}