+++ packages/sdks/output/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Builder.io React SDK Changelog (@builder.io/sdk-react)
 
+## 0.13.1
+
+## 0.13.0
+
+### Minor Changes
+
+- da5d871: 🧨 Breaking Change: remove 'v2' as a viable `apiVersion`. Only 'v3' is now allowed.
+
 ## 0.12.8
 
 ### Patch Changes